fix(cap): unblock Capacitor build by extracting nutrition types out of /app/api
Three nutrition route files exported types (HistoryDay, SummaryDay,
FoodResult) that were being type-imported from client pages and lib
helpers. The Capacitor build (build:cap) moves src/app/api out of the
tree before next build to disable API routes, which broke those
type-only imports and failed the build with "Cannot find module
@/app/api/nutrition/...".

Moved the three interfaces to src/lib/nutrition-history-types.ts. The
route files re-export from there so existing server callers keep working,
and client code now imports from @/lib/... — independent of api/.

Also added -allowProvisioningDeviceRegistration to ios-device-build.sh
so a freshly-paired device gets auto-registered with App Store Connect
via the API key (instead of failing with "No provider was found" the
first time).
